temporary file management
This commit updates how Parsons handles temporary files. The
changes are meant to minimize the need for users to manage temp
files themselves, specifically when using a Parsons `Table`.

This commit introduces a new class
`parsons.utilities.resources.ResourceManager` that tracks
resources that are used by one or more `Table` so that the
resources (i.e. temp files) are cleaned up when they are no
longer being used by the table(s).