google cloud storage: don't get bucket before dling blob (#120)
This commit updates the `GoogleCloudStorage` connector to not call
`get_bucket` when downloading a blob to a local file. It is
possible in Cloud Storage to have access to a blob in a bucket
without actually having access to get the bucket metadata (which
is what `get_bucket` does). The underlying Cloud Storage Python
client has a `download_blob_to_file` method which does not require
that we call `get_bucket`, so this commit switches the Parsons
`GoogleCloudStorage`'s `download_blob` method to use that.