honor delimiter on redshift copy_s3
This commit passes through the delimiter to the `from_csv` method
used to load a CSV file when copying from S3 to Redshift. The call
in question is part of Parsons' effort to reflect on the data
being copied to build a `create table` statement in the case where
the target Redshift table does not exist.