Add ability to authenticate using a private rsa key over sftp (#139)
This commit provides the option to authenticate an sftp connection using a private rsa key as an alternative to username and password authenticate. If a password is provided in addition to a key file path, both will be passed to the paramiko sftp connection configuration. I've added some tests, but they are tests that require a live test environment to succeed.