document low resource patterns (#157)
This commit adds a section to the Parsons docs on utilizing
Parsons in environments where resource consumption is a concern.
The goal is to allow developers to avoid loading all of the
Parsons connectors (and dependent libraries) whenever loading
any of them. In order to make this possible, this commit
introduces an environment variable `PARSONS_SKIP_IMPORT_ALL`
that can be set to skip over importing all classes in the
root `parsons` package.