add noise to upsert staging table name
This commit changes the naming convention for the staging table
used by the `Redshift.upsert` method. This commit adds `stg` to
the name of the table. It also adds some random noise to the
end of the table. Before this commit, running the `upsert` command
in parallel against the same table could result in errors from
Redshift from trying to create the same table at the same time.