etl: convert_columns_to_str handle no types (#212)
This commit updates the `convert_columns_to_str` method in the
Parsons `Table` handle a case where `petl` was returning an
empty list for the types from the `typeset` function. The
Parsons code had assumed that the list returned from `typeset`
would always be non-empty, so when an empty value was returned
the code raised an index out of bounds exception.

It looks like the reason `typeset` was returning an empty list
of types is that the `Table` in question didn't have any rows,
so this commit exits early from the function if there's no data.
Also, to be safe, the commit updates the conditional check to
short circuit if there is less than one (no types) or more than
one (multiple types) value in the list.