bigquery: DB sync
This commit updates the Parsons BigQuery connector to support DB
syncs to/from BigQuery. In order to make this change, the commit
also modifies the `query` method in the connector to use the
BigQuery
[DB-API](https://googleapis.dev/python/bigquery/latest/dbapi.html)
interface.

The commit also updates the BigQuery connector API to
require users to qualify their table names with the name of the
dataset, so align the BigQuery API with the more standard SQL
connectors (Redshift, MySQL).

This commit also makes a few minor changes to the DB sync class
to fix some small bugs, namely:
 * Add the `table` method to MySQL to allow using MySQL for with
 `DBSync`
 * Ensure that the connection and cursor in the Redshift connector
 are closed, even if the query raises an exception.
 * Fixes a bug where the DB sync API wouldn't support incremental
 table syncs where the primary key wasn't an integer.
 * Work around a limitation in BigQuery where the `autodetect`
 of a copy job would confuse headers for a data row in the case
 where the table data was all strings.
 * Fix DB Sync `table_sync_incremental` so it works even if the
 destination table is empty.