phone2action: handle empty advocates results
This commit updates the `Phone2Action` connector to handle an
empty list from the call to the advocates endpoint on the P2A API.
Before the commit, the call to get advocates was failing when
trying to unpack the advocates data.

This commit also adds the `__bool__` magic method to the Parsons
`Table` class to support simple bool checks on the emptiness of
a table (e.g. `if not tbl:` instead of `if tbl.num_rows > 0`)