materialize_to_file method (#264)
This commit adds the `materialize_to_file` method, which allows
a user to apply any pending transformations to a table. The table
data is written to a temp file, which overcomes a limitation of
the current `materialize` method, which requires loading all of
the data from the table into memory.