Redshift upsert: auto resize column widths if necessary (#273)
Resizes the target table columns in Redshift based on the incoming Parsons table before creating the staging table in Redshift, so that when the staging table is created it already is at the correct column widths. This also means that an upsert will sometimes error out earlier if the incoming table doesn't match the columns of the target table, a side effect that could save some processing time on failed upserts.