docs: update quickstart
This commit updates the Quick Start code in the README.md file
for Parsons so that the quick start can be executed without
credentials. The goal is to make it so that anyone could run
the Quick Start code to begin understanding how to use Parsons.

This commit also adds a `download_table` method to the `GitHub`
Connector that allows downloading CSV data from GitHub as a
Parsons `Table`.

Finally, this commit tweaks the `download_file` method on the
`GitHub` connector to support token-based authentication when
downloading files / tables.