actionkit: use `with` when uploading
This commit updates the ActionKit Connector to use the `with`
statement when using `requests` to upload CSV data. This commit
is meant to handle an error seen by a user where connections
weren't properly closed after upload. The `with` statement should
ensure that `requests` properly closes connections between
calls to the ActionKit API.