phone2action: add update_advocate / create_advocate
This commit adds the `update_advocate` and `create_advocate`
methods to the `Phone2Action` Connector class.

This commit also updates the `APIConnector` class to allow users
to only have to pass in the relative path to an endpoint on
request calls to simplify the API. It also updates any classes
that were passing in the full URL to only pass in the relative
path. *Note:* It is still possible to pass in a full URL; the
full URL will override the base URL in the connector.

Finally, this commit tweaks the existing date utilities by
a) adding a new `parse_date` function for parsing semi-arbitrary
date values, and b) to rework the `iso_to_unix` function to the
`date_to_timestamp` function that can take any date-like value
and convert it into a UNIX timestamp.