redshift: don't overfit varchar length
This commit updates the `RedshiftCreateTable` class to use a step
function to determine the best varchar length instead of using
just the length of the longest string. The goal of this PR is to
reduce the need for future column resizes when more data is
loaded, while also keeping the column sizing as small as possible.
This commit also removes the use of `smallint` as a possibility
for an integer column value.