table: improve __bool__ performance
This commit updates the `__bool__` magic method implementation
in the Parsons `Table` to be more performant. Before this PR the
`__bool__` method used the `num_rows` to determine whether or not
there was any data in the `Table`. The problem is that `num_rows`
will scan the whole `Table`, but `__bool__` only needs to know if
there is a single row of data. The new implementation leverages
the `petl.head` method to pull out a single row from the `Table`.