dbsync: improve performance
This commit improves performance of the `DBSync` class by making
distinct checks and row counts optional. The commit is meant to
allow users to opt out of these features to workaround issues
seen with really big tables.

This commit also fixes a bug from a previous commit where an
invalid class attribute was being accessed. The value is now
pulled in via a constant in another module.