Update docs to account for wheels. (#698)
Parsons is hosted on pypi as both a source distribution and a binary
wheel package. The later is faster, but skips over configurations in
setup.py (more specifically: whatever environment vars were set on the
building machine are the ones that stuck). To allow
`PARSONS_LIMITED_DEPENDENCIES` to work, we need to tell pip that it
should skip the wheel for Parsons.