Don't auto-grab session token if it's not passed in (#740)
* Don't auto-grab session token if it's not passed in

It turns out that trying to grab a session token from the environment variable can break things when used in concert with Zappa asynchronous processing - the access key and secret key passed in may not match the temporary token generated for the asynchronous process. Thus, this change requires the token to be passed in, even if it's set as an environment variable. 

As described [here](https://github.com/zappa/Zappa#asynchronous-task-execution), with asynchronous task execution, Zappa spins up a completely separate Lambda instance. The session token is generated as part of that process, and (as far as I can tell) is based on the execution role of the Lambda function.
So when the asynchronous task is running, it has the `aws_session_token` set even though there wasn't one originally. I'm not sure why it isn't just "None".
In any case, the token doesn't match the access key ID and secret access key that I was being passed in, resulting in an InvalidTokenError: `An error occurred (InvalidToken) when calling the PutObject operation: The provided token is malformed or otherwise invalid`. It's not that the token is actually invalid, but it doesn't match the credentials being passed in.

I'd like to be able to force aws_session_token to None even if the environment variable is set.

* Add optional argument to ignore env token

* Spacing