Fix match_columns
The function definition of match_column states that if there are 2 columns with the same normalized name, the latter should be removed by default. The normalize_fn function must run over the reversed list of self.columns in order for the first matching column to be kept. In it's previous state, the last matching column was kept