Fix assert_matching_tables (#759)
* Fix tests

* Only use .items() to assert matching tables for dicts

* Fix test_aws_async tests

The csv_renewed and distribute_task functions return a table with the ints converted to strs

* Fix match_columns

The function definition of match_column states that if there are 2 columns with the same normalized name, the latter should be removed by default. The normalize_fn function must run over the reversed list of self.columns in order for the first matching column to be kept. In it's previous state, the last matching column was kept

* Fix order of records

To match order in airtable_responses.py

* Fix Hustle create_leads test

Current test is returning the Lyndon Johnson lead twice

* Fix test_targets.py

Fix typo in test. Also change import unittest to import unittest.mock to avoid `AttributeError: module 'unittest' has no attribute 'mock'` when running test locally

* Figure out if r1 and r2 are lists separately