Refactor petl table validation on Table init to avoid iteration
Very large tables read into Table.from_csv will take a long time to
initialize. This is due to a validation method that is meant to check
for a valid underlying petl Table.

Checking the truthiness of a petl Table defaults to checking its
length, which iterates one by one over every row in the underlying
data. For a very long file, this takes a very long time. It is
unnecessary and also doesn't confirm or validate what we want to
validate.

This new implementation explicitly checks that the underlying table is
a petl Table.