Refactor Table initialization to avoid iterating over every row (#983)
* Refactor petl table validation on Table init to avoid iteration

Very large tables read into Table.from_csv will take a long time to
initialize. This is due to a validation method that is meant to check
for a valid underlying petl Table.

Checking the truthiness of a petl Table defaults to checking its
length, which iterates one by one over every row in the underlying
data. For a very long file, this takes a very long time. It is
unnecessary and also doesn't confirm or validate what we want to
validate.

This new implementation explicitly checks that the underlying table is
a petl Table.

* Refactor Table initialization for robustness

A few changes are made:

- replace mutable default argument

See https://docs.python-guide.org/writing/gotchas/#mutable-default-arguments
for why a mutable variable should not be used as a default argument

- Replace calls to type() with isinstance()

isinstance() is generally recommended over comparisons to type()
because it is faster and considers inheritance

- Explicitly check if the argument passed is a petl Table

Previous control flow would assume that the argument works and then
run a validation check on it afterward, but this method checks
explicitly.

* Use sentinal as default argument to maintain existing behavior

Existing behavior allows Table() to work while failing on Table(None)

This requires not using None as a default argument but instead using
an empty sentinal

* Fix NGPVAN test mocker responses to use lists of dicts, not a dict

This change brings the mocked responsese into alignment with what is
actually returned by the underlying methods

* Fix formatting