Allow ActionNetwork upsert_person to not change mobile opt-in status (#1212)
Prior behavior opted-in all phone numbers by default, and the only
alternative would be to opt-out phone numbers sent to the API. This
update allows for not changing the opt-in status of the number, and
changes the default from opt-in to no change. This is a more
reasonable and legally compliant default behavior than opting-in.