Consolidate and DRY up tests on DB Sync utility (#1242)
Consolidates duplicative code into a standardized TestDBSync abstract base class, which can then be inherited and run the same tests on the FakeDatabase, Postgres, and Redshift connectors. Technically this new setup extends the test surface a bit, especially since a full suite of tests now runs on the FakeDatabase in the CICD environment (without LIVE_TEST=1). Previously only a very limited set of tests ran on the FakeDatabase in the CICD environment.